move targetresolver to controller-utils package (run-int-tests)
