+++ pkg/landscaper/installations/executions/template/gotemplate/funcs.go
@@ -24,11 +24,11 @@ import (
 	"sigs.k8s.io/yaml"
 
 	"github.com/gardener/landscaper/apis/core/v1alpha1"
+	"github.com/gardener/landscaper/controller-utils/pkg/landscaper/targetresolver"
 	"github.com/gardener/landscaper/pkg/components/model"
 	"github.com/gardener/landscaper/pkg/components/model/types"
 	lstmpl "github.com/gardener/landscaper/pkg/landscaper/installations/executions/template"
 	"github.com/gardener/landscaper/pkg/utils/clusters"
-	"github.com/gardener/landscaper/pkg/utils/targetresolver"
 )
 
 // LandscaperSprigFuncMap returns the sanitized spring function map.
