+++ pkg/utils/clusters/shoot_kubeconfigs.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/gardener/landscaper/apis/core/v1alpha1"
 	"github.com/gardener/landscaper/apis/core/v1alpha1/targettypes"
-	"github.com/gardener/landscaper/pkg/utils/targetresolver"
+	"github.com/gardener/landscaper/controller-utils/pkg/landscaper/targetresolver"
 )
 
 // GetShootClusterNameFromKubeconfig determines the name of a Gardener shoot cluster from a given kubeconfig.
